==> 00-senadores.Rmd <==
---
title: "Senadores"
author: "Bernardino Araya"
date: "2025-11-06"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r}
library(tidyverse)
library(writexl)
library(readxl)
```

```{r}
bd_sen_excel <- read_excel("Bases de datos/2021_11_Senadores_Datos_Eleccion.xlsx", skip = 4)
```

### Paso 1 
Identifiqué los nombres de las variables om columnas usando la función **colnames()**

```{r}
colnames(bd_sen_excel)
```

### Paso 2
Como muchos de los nombres no me servian (eran cosas como ...2, ...5, ...8, etc), usé la función colnames() y c() para reasignarle nombre a cada variable.

```{r}
colnames(bd_sen_excel) <- c("nro_region", 
                            "region", 
                            "circ_sen", 
                            "distrito", 
                            "comuna", 
                            "lista", 
                            "pacto", 
                            "partido", 
                            "nro_voto", 
                            "nombres", 
                            "primer_apellido", 
                            "segundo_apellido", 
                            "votos", 
                            "cargo"
                            )
```

### Paso 3
Como habia tres variables (nombres, primer_apellido y segundo_apellido) que funcionaban como categorias distintas, las fucione bajo el nombre de "candidato" con el orden: primer_apellido segundo_apellido, nombres.

```{r}
bd_sen_excel <- bd_sen_excel |> select(region, 
                                       circ_sen,
                                       distrito, 
                                       lista, 
                                       pacto,
                                       partido, 
                                       nro_voto,
                                       nombres,
                                       primer_apellido, 
                                       segundo_apellido, 
                                       votos, 
                                       cargo
                                       ) |>
 mutate(candidato = paste(nombres, primer_apellido, segundo_apellido))
```

```{r}
bd_sen_excel <- bd_sen_excel |>  group_by(region,
                                          circ_sen,
                                          lista, 
                                          partido, 
                                          nro_voto, 
                                          candidato, 
                                          cargo, 
                                          pacto, 
                                          ) |>
  summarise(total_region = sum(votos, na.rm = TRUE)) |> 
  filter(!is.na(partido))
```

```{r}
senadores <- bd_sen_excel |> select(region, 
                                    circ_sen, 
                                    lista, 
                                    nro_voto, 
                                    pacto, 
                                    partido, 
                                    candidato, 
                                    total_region, 
                                    cargo
                                    )
```

Guardar base de dato:

```{r}
write_rds(senadores, "output/base_senadores.rds")
```
